chore: relocate mocks into their own package
This should keep the mocks out of the coverage map (which is good)
and the production binary build (which is great)
