+++ Taskfile.yml
@@ -115,7 +115,7 @@ tasks:
   mockgen:
     desc: Generate mock for interface
     cmds:
-      - mockgen -source=internal/git.go -destination=internal/git/git_mock.go -package=git
-      - mockgen -source=internal/repository.go -destination=internal/repository/repository_mock.go -package=repository
-      - mockgen -source=internal/exec.go -destination=internal/exec/exec_mock.go -package=exec
-      - mockgen -source=internal/repository/types.go -destination=internal/repository/copy_mock.go -package=repository
+      - mockgen -source=internal/git.go -destination=internal/mocks/git/git_mock.go -package=git
+      - mockgen -source=internal/repository.go -destination=internal/mocks/repository/repository_mock.go -package=repository
+      - mockgen -source=internal/exec.go -destination=internal/mocks/exec/exec_mock.go -package=exec
+      - mockgen -source=internal/repository/types.go -destination=internal/mocks/repository/copy_mock.go -package=repository
