+++ examples/go-client/go.mod
@@ -15,7 +15,6 @@ require (
 	github.com/go-playground/locales v0.14.1 // indirect
 	github.com/go-playground/universal-translator v0.18.1 // indirect
 	github.com/go-playground/validator/v10 v10.16.0 // indirect
-	github.com/golang/mock v1.6.0 // indirect
 	github.com/leodido/go-urn v1.2.4 // indirect
 	github.com/spf13/afero v1.11.0 // indirect
 	golang.org/x/crypto v0.16.0 // indirect
