+++ internal/repositories/repositories_public_test.go
@@ -33,9 +33,9 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	"github.com/retr0h/go-gilt/internal"
-	"github.com/retr0h/go-gilt/internal/exec"
+	"github.com/retr0h/go-gilt/internal/mocks/exec"
+	"github.com/retr0h/go-gilt/internal/mocks/repository"
 	"github.com/retr0h/go-gilt/internal/repositories"
-	"github.com/retr0h/go-gilt/internal/repository"
 	"github.com/retr0h/go-gilt/pkg/config"
 )
