+++ internal/repositories/repositories_test.go
@@ -30,8 +30,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/retr0h/go-gilt/internal/exec"
-	"github.com/retr0h/go-gilt/internal/repository"
+	"github.com/retr0h/go-gilt/internal/mocks/exec"
+	"github.com/retr0h/go-gilt/internal/mocks/repository"
 	"github.com/retr0h/go-gilt/pkg/config"
 )
